Sherman
Bring the good old bugle, boys, we'll sing another song
Sing it with a spirit that will start the world along
Sing it as we used to sing it, 50, 000 strong
While we were marchin' through Georgia
Hurrah! Hurrah! We bring the jubilee
Hurrah! Hurrah! The flag that makes you free
So we sang the chorus from Atlanta to the sea
While we were marchin' through Georgia
How the people shouted when they heard the joyful sound?
How the turkeys gobbled which our commissary found?
How the sweet potatoes even started from the ground?
While we were marchin' through Georgia
Yes and there were Union men who wept with joyful tears
When they saw the honored flag they had not seen for years
Hardly could they be restrained from breaking forth in cheers
While we were marchin' through Georgia
Hurrah! Hurrah! We bring the jubilee
Hurrah! Hurrah! The flag that makes you free
So we sang the chorus from Atlanta to the sea
While we were marchin' through Georgia
Sherman's dashin' Yankee boys will never reach the coast!
So the saucy rebels said and 'twas a handsome boast
Had they not forgot, alas! To reckon with the Host
While we were marchin' through Georgia
So we made a thorough fare for freedom and her train
60 miles in latitude, 300 to the main
Treason fled before us, for resistance was in vain
While we were marchin' through Georgia
Hurrah! Hurrah! We bring the jubilee
Hurrah! Hurrah! The flag that makes you free
So we sang the chorus from Atlanta to the sea
While we were marchin' through Georgia
Hurrah! Hurrah! We bring the jubilee
Hurrah! Hurrah! The flag that makes you free
So we sang the chorus from Atlanta to the sea
While we were marchin' through Georgia
While we were marchin' through Georgia
While we were marchin' through Georgia